+++ XMapper.Testing.Tests/DummyAssemblyWithStaticFieldTests.cs
@@ -1,14 +1,22 @@
 ﻿using System;
 using Xunit;
+using Xunit.Abstractions;
 
 namespace XMapper.Testing.Tests;
 public class DummyAssemblyWithStaticFieldTests
 {
+    private readonly XMapperValidator _validator;
+
+    public DummyAssemblyWithStaticFieldTests(ITestOutputHelper testOutputHelper)
+    {
+        _validator = new XMapperValidator(testOutputHelper.WriteLine);
+    }
+
     [Fact]
     public void PropertyNotFound()
     {
         var exception = Assert.ThrowsAny<Exception>(() =>
-            AssertXMapper.AllAreValidInAssembly("DummyAssembly1", TestCases.AppDefaults));
+            _validator.AllAreValidInAssembly("DummyAssembly1", TestCases.AppDefaults));
         Assert.Contains("DummyAssembly1.Class1.MapperField", exception.Message);
         Assert.Contains("Property 'XString' was not found on source 'DummyA'.", exception.Message);
     }
@@ -17,7 +25,7 @@ public class DummyAssemblyWithStaticFieldTests
     public void MismatchingIgnoreWithPropertyList()
     {
         var exception = Assert.ThrowsAny<Exception>(() =>
-            AssertXMapper.AllAreValidInAssembly("DummyAssembly3", TestCases.AppDefaults));
+            _validator.AllAreValidInAssembly("DummyAssembly3", TestCases.AppDefaults));
         Assert.Contains("DummyAssembly3.Class1.MapperField", exception.Message);
         Assert.Contains("Use 'IgnoreSourceProperty' if PropertyList is Source.", exception.Message);
     }