+++ XMapper.Testing.Tests/TestHelpers/Does_NotThrow.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace XMapper.Testing.Tests;
+
+public static class Does
+{
+    public static void NotThrow(Action action)
+    {
+        action();
+    }
+}